+++ sdk/nodejs/getTemplate.ts
@@ -30,7 +30,7 @@ export interface GetTemplateArgs {
      */
     readonly filters?: inputs.GetTemplateFilter[];
     /**
-     * If is used, them all instances will be from that region.
+     * If is used, them all template will be from that region, has to be declared here if is not declared in the provider
      */
     readonly region?: string;
     /**